pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "pyflyde"
version = "0.0.1"
requires-python = ">= 3.9"

dependencies = [
    "pyyaml",
]
